Move ParseExtrVars to main.go.
It was only being used in main, so it didn't make much sense to keep in
utils.
