+++ claat/cmd/util.go
@@ -20,8 +20,6 @@
 package cmd
 
 import (
-	"encoding/json"
-	"log"
 
 	// allow parsers to register themselves
 	_ "github.com/googlecodelabs/tools/claat/parser/gdoc"
@@ -42,25 +40,7 @@ const (
 	reportOk  = "ok\t%s"
 )
 
-var ExtraVars map[string]string // Extra template variables passed on the command line.
-
 // isStdout reports whether filename is stdout.
 func isStdout(filename string) bool {
 	return filename == stdout
-}
-
-// ParseExtraVars parses extra template variables from command line.
-// extra is any additional arguments to pass to format templates. Should be formatted as JSON objects of string:string KV pairs.
-func ParseExtraVars(extra string) (map[string]string, error) {
-	vars := make(map[string]string)
-	if extra == "" {
-		return vars, nil
-	}
-	b := []byte(extra)
-	err := json.Unmarshal(b, &vars)
-	if err != nil {
-		log.Printf("Error parsing additional template data: %v", err)
-		return nil, err
-	}
-	return vars, nil
 }
